requirements.txt
pyyaml>=6.0
pdfplumber>=0.10
thefuzz>=0.22
python-Levenshtein>=0.25    # speeds up thefuzz
openpyxl>=3.1
requests>=2.31
google-auth>=2.28
google-auth-oauthlib>=1.2
google-api-python-client>=2.120
pdf2image>=1.17             # OCR fallback — also requires Tesseract + Poppler
pytesseract>=0.3            # OCR fallback
